[refactor] separate layout constants
